LayoutTests/platform/chromium/fast/forms/datetimelocal/datetimelocal-stepup-stepdown-from-renderer-expected.txt

Check stepping-up and -down for datetime-local input fields from renderer. No cases of empty initial values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Function arguments are (value, step, {min or max}, [stepCount]).
Normal cases
FAIL stepUp("2010-02-10T20:13", null, null) should be 2010-02-10T20:14. Was 2010-02-10T20:13.
FAIL stepDown("2010-02-10T20:13", null, null) should be 2010-02-10T20:12. Was 2010-02-10T20:13.
FAIL stepUp("2010-02-10T20:13", null, null, 10) should be 2010-02-10T20:23. Was 2010-02-10T20:13.
FAIL stepDown("2010-02-10T20:13", null, null, 11) should be 2010-02-10T20:02. Was 2010-02-10T20:13.
FAIL stepUp("1970-01-01T20:13", "4", null, 2) should be 1970-01-01T20:13:08. Was 1970-01-01T20:13.
FAIL stepDown("1970-01-01T20:13", "4", null, 3) should be 1970-01-01T20:12:48. Was 1970-01-01T20:13.
Step=any
FAIL stepUp("2010-02-10T20:13", "any", null) should be 2010-02-10T20:14. Was 2010-02-10T20:13.
FAIL stepDown("2010-02-10T20:13", "any", null) should be 2010-02-10T20:12. Was 2010-02-10T20:13.
Overflow/underflow
FAIL stepUp("2010-02-10T20:13", "3.40282346e+38", null) should be 275760-09-13T00:00:00. Was 2010-02-10T20:13.
FAIL stepDown("2010-02-10T20:13", "3.40282346e+38", null) should be 1970-01-01T00:00:00. Was 2010-02-10T20:13.
PASS stepUp("2010-02-10T20:13", "1", "2010-02-10T20:13") is "2010-02-10T20:13"
PASS stepDown("2010-02-10T20:13", "1", "2010-02-10T20:13") is "2010-02-10T20:13"
stepDown()/stepUp() for stepMismatch values
FAIL stepDown("2010-02-10T20:13", "3", "2010-02-10T20:12:56") should be 2010-02-10T20:12:59. Was 2010-02-10T20:13.
FAIL stepUp("1970-01-01T00:13", "7", "") should be 1970-01-01T00:13:04. Was 1970-01-01T00:13.

PASS successfullyParsed is true

TEST COMPLETE
